updating after a long time

+++ offline_tools/processing/cnn_application.py
@@ -5,9 +5,14 @@ import os
 import time
 from shutil import copyfile
 
-epoch_path = msc.joinToHome("/data/epochs/3321/model.ckpt")
+gen_extras = True
 
-out_basepath = msc.joinToHome("/data/cnn_output")
+# epoch_path = msc.joinToHome("/data/epochs/3321/model.ckpt")
+epoch_path = msc.joinToHome("data/epochs/checkpoints3/1033/model.ckpt")
+
+
+# out_basepath = msc.joinToHome("/data/cnn_output")
+out_basepath = msc.joinToHome("data/cnn_output3")
 
 out_dirs = ['orig_resiz','overlay','masks','ndvi','only_veg']
 
@@ -20,8 +25,16 @@ for name in out_dirs:
 
 
 imgs_path = "/home/kaue/data/extracted_images/2019-07-11-16-21-46/ngr"
+# imgs_path = "/home/kaue/data/extracted_images/red_band"
 
-img_list = msc.orderedFileList(imgs_path)
+
+img_list = msc.orderedFileList(imgs_path,extension='*.jpg')
+
+img_list = sorted(img_list)
+
+# del img_list[0:225]
+
+print(img_list[0:5])
 
 for i,img_path in enumerate(img_list):
     t1 = time.time()
@@ -39,15 +52,19 @@ for i,img_path in enumerate(img_list):
     dest_overlay = os.path.join(outpaths['overlay'],img_name+'.png')
 
     try:
-        vd.single_img_forward_pass(epoch_path,img_path)
+        if not os.path.isfile(dest_pred):
+            vd.single_img_forward_pass(epoch_path,img_path)
 
-        copyfile(out_orig,dest_orig)
-        copyfile(out_pred,dest_pred)
+            # vd.single_img_forward_pass(epoch_path,img_path,dataset='cityscapes',model='DeepLabV3_plus')
 
-        cs.gen_only_veg_img(dest_pred,dest_orig,dest_onlyveg)
-        cs.gen_uint8_NDVI(dest_onlyveg,dest_ndvi)
+            copyfile(out_orig,dest_orig)
+            copyfile(out_pred,dest_pred)
+        
+        if gen_extras:
+            cs.gen_only_veg_img(dest_pred,dest_orig,dest_onlyveg)
+            cs.gen_uint8_NDVI(dest_onlyveg,dest_ndvi)
 
-        cs.gen_overlay_img(dest_orig,dest_pred,dest_overlay)
+            cs.gen_overlay_img(dest_orig,dest_pred,dest_overlay)
 
 
     except:
@@ -57,4 +74,6 @@ for i,img_path in enumerate(img_list):
     msc.print_rem_time_info(len(img_list),i,t1)
 
 
-# msc.telegram_bot_sendtext("img_gen done!!")
+msc.telegram_bot_sendtext("img_gen done!!")
+
+# cityscapes dict: /home/kaue/Semantic-Segmentation-Suite/cityscapes/class_dict.csv
